refactor(ww3d2): move backend implementations into Backend subfolder
