+++ Core/Libraries/Source/WWVegas/WW3D2/Backend/DX8Backend.cpp
